Add Linux specific target
Create a target that is Linux specific. Linux needs a dramatically
different feature set as to what is already in riscv-opcodes. Primarily,
it needs the ability to create and modify entire instructions, including
immediates.

Immediates were not previously represented in riscv-opcodes so
variable_field_data.py was created.

Linux would also prefer to have a config file that can be passed it to
specify which instructions are supported.

Having a more advanced argument system like argparse makes it easier to
specify options like the path for the output file and config file, and
that was not possible in the current parse.py file.

Signed-off-by: Charlie Jenkins <charlie@rivosinc.com>